+++ src/main/java/util/GameUtils.java
@@ -5,7 +5,7 @@ import model.game.player.Player;
 import java.util.List;
 import java.util.function.BinaryOperator;
 
-public class GameUtil {
+public class GameUtils {
     private static final BinaryOperator<String> compareString = (s1, s2) -> s1.getBytes().length >= s2.getBytes().length ? s1 : s2;
 
     public static int getMaxPlayerName(List<Player> players) {
