+++ src/main/java/util/StringUtils.java
@@ -1,6 +1,6 @@
 package util;
 
-public class StringUtil {
+public class StringUtils {
     public static String rpad(int pad, String player) {
         return String.format("%-" + pad + "s", player);
     }
